Added attributes to blocks and tables and changed web page look
It is now possible to add attributes to source code blocks.

Code blocks that are about command-line languages have the attributes
USERNAME, HOSTNAME and HIGHLIGHT. It is possible to add the username and
hostname of the command and it is possible to highlight the specified lines.
The output on the command-line source code block is specified by prefixing
the code with '(out)'.

Source code blocks with other programming languages have the attribute HIGHLIGHT
to highlight specified lines, the attribute FETCH to get files from the Github
API, the attribute FILEPATH to specify a file in a remote path and add a
download button to the code block in HTML

The special code blocks TIP, WARNING and DANGER have the attribute NAME
to give a title to the block.

Tables have the attribute NAME to add a description of the table.

All the headers bellow level 6 will have the same style.