Updated content and changed dark-theme settings
Changed the project settings as it didn't include the base extension of the files.
This would lead to missing files in the dist directory.

Added content to the table to make it easier to see the column alignment.

Added an author to the blockquote.

Updated the attributes description list.

Changed the dark-mode settings to a cleaner theme.