Changed theme, fixed TOC and added scroll spy
Added a new package as dependency to allow easier string manipulation.

Changed theme to make it cleaner.

Removed CSS from the HTML template and added it to the styles.css file.

Added scroll spy to show which chapter is currently being displayed.

Added a author attribute to the blockquote block.

Added column alignment for tables. In the org file you just have to add `<l|c|r>` for
left, center or right alignment respectively.

Fixed TOC json export to allow changing the nesting level of the headers. It is based on
the `org-tailwind-headlines` variable and it counts the commas + 1 to find the desired
heading level.